+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 All notable changes to this project will be documented in this file.
 
-## NEXT
+## 1.3.1 (2020-11-22)
 
 * Prevent garage door open/close command if currently in opening/closing state
 * Support brightness level for Omni units